Updated the filter for Washington Posts
WashingtonPosts introduce a version of a paywall when a class it is
marked with is autogenerated. It's quite difficult to find the paywall
block using regular CSS.

To get around this problem xpath should fit pretty well. See the link
to the SO answer below.

However, it's not enough to remove the paywall block, since they
also make the page itself not scrollable by assigning overflow
attribute to the html node. We need to replace hidden by fix.

To make sure that paywall isn't removed, WashingtonPost website
doesn't display the wall for some time (maybe it takes some time to load).
That's why we have to try several times until the wall appears.
I used 10 attempts 500ms, it should give us 5s to block the wall.

SO xpath answer: https://stackoverflow.com/a/29289196/2632860